+++ framework_src/FreSharp/FreSharpHelper.cs
@@ -118,12 +118,7 @@ namespace TuaRua.FreSharp {
         }
 
         internal static uint GetArgsC(ArrayList args) {
-            uint cnt = 0;
-            if (args != null) {
-                cnt = (uint) args.Count;
-            }
-
-            return cnt;
+            return args != null ? (uint) args.Count : 0;
         }
 
         internal static FREObject NewObject(string value) {
@@ -131,7 +126,7 @@ namespace TuaRua.FreSharp {
             var ret = Core.getFREObject(value, ref resultPtr);
             var status = (FreResultSharp) resultPtr;
             if (status == FreResultSharp.Ok) return ret;
-            Logger.Log($"cannot create FREObject from {value}", status);
+            Logger.Error($"cannot create FREObject from {value}", status);
             return FREObject.Zero;
         }
 
@@ -144,7 +139,7 @@ namespace TuaRua.FreSharp {
             var ret = Core.getFREObject(value, ref resultPtr);
             var status = (FreResultSharp) resultPtr;
             if (status == FreResultSharp.Ok) return ret;
-            Logger.Log($"cannot create FREObject from {value}", status);
+            Logger.Error($"cannot create FREObject from {value}", status);
             return FREObject.Zero;
         }
 
@@ -153,7 +148,7 @@ namespace TuaRua.FreSharp {
             var ret = Core.getFREObject(value, ref resultPtr);
             var status = (FreResultSharp) resultPtr;
             if (status == FreResultSharp.Ok) return ret;
-            Logger.Log($"cannot create FREObject from {value}", status);
+            Logger.Error($"cannot create FREObject from {value}", status);
             return FREObject.Zero;
         }
 
@@ -162,7 +157,7 @@ namespace TuaRua.FreSharp {
             var ret = Core.getFREObject(value, ref resultPtr);
             var status = (FreResultSharp) resultPtr;
             if (status == FreResultSharp.Ok) return ret;
-            Logger.Log($"cannot create FREObject from {value}", status);
+            Logger.Error($"cannot create FREObject from {value}", status);
             return FREObject.Zero;
         }
 
@@ -171,7 +166,7 @@ namespace TuaRua.FreSharp {
             var ret = Core.getFREObject(value, ref resultPtr);
             var status = (FreResultSharp) resultPtr;
             if (status == FreResultSharp.Ok) return ret;
-            Logger.Log($"cannot create FREObject from {value}", status);
+            Logger.Error($"cannot create FREObject from {value}", status);
             return FREObject.Zero;
         }
 
@@ -180,7 +175,7 @@ namespace TuaRua.FreSharp {
             var ret = Core.getString(rawValue, ref resultPtr);
             var status = (FreResultSharp) resultPtr;
             if (status == FreResultSharp.Ok) return StringUtils.ToUtf8(ret);
-            Logger.Log($"cannot get FREObject {rawValue.toString()} as String", status);
+            Logger.Error($"cannot get FREObject {rawValue.toString()} as String", status);
             return null;
         }
 
@@ -189,7 +184,7 @@ namespace TuaRua.FreSharp {
             var ret = Core.getDouble(rawValue, ref resultPtr);
             var status = (FreResultSharp) resultPtr;
             if (status == FreResultSharp.Ok) return ret;
-            Logger.Log($"cannot get FREObject {rawValue.toString()} as Double", status);
+            Logger.Error($"cannot get FREObject {rawValue.toString()} as Double", status);
             return 0.0;
         }
 
@@ -198,7 +193,7 @@ namespace TuaRua.FreSharp {
             var ret = Core.getBool(rawValue, ref resultPtr);
             var status = (FreResultSharp) resultPtr;
             if (status == FreResultSharp.Ok) return ret;
-            Logger.Log($"cannot get FREObject {rawValue.toString()} as Bool", status);
+            Logger.Error($"cannot get FREObject {rawValue.toString()} as Bool", status);
             return false;
         }
 
@@ -207,7 +202,7 @@ namespace TuaRua.FreSharp {
             var ret = Core.getInt32(rawValue, ref resultPtr);
             var status = (FreResultSharp) resultPtr;
             if (status == FreResultSharp.Ok) return ret;
-            Logger.Log($"cannot get FREObject {rawValue.toString()} as Int", status);
+            Logger.Error($"cannot get FREObject {rawValue.toString()} as Int", status);
             return 0;
         }
 
@@ -216,7 +211,7 @@ namespace TuaRua.FreSharp {
             var ret = Core.getUInt32(rawValue, ref resultPtr);
             var status = (FreResultSharp) resultPtr;
             if (status == FreResultSharp.Ok) return ret;
-            Logger.Log($"cannot get {rawValue.toString()} as UInt", status);
+            Logger.Error($"cannot get {rawValue.toString()} as UInt", status);
             return 0;
         }
 
@@ -252,7 +247,7 @@ namespace TuaRua.FreSharp {
                 ref resultPtr);
             var status = (FreResultSharp) resultPtr;
             if (status == FreResultSharp.Ok) return;
-            Logger.Log($"cannot set property {name} to {rawValue.toString()}", status, ret);
+            Logger.Error($"cannot set property {name} to {rawValue.toString()}", status, ret);
         }
 
         internal static void SetProperty(FREObject rawValue, string name, FREObject value) {
@@ -260,7 +255,7 @@ namespace TuaRua.FreSharp {
             var ret = Core.setProperty(rawValue, name, value, ref resultPtr);
             var status = (FreResultSharp) resultPtr;
             if (status == FreResultSharp.Ok) return;
-            Logger.Log($"cannot set property {name} to {rawValue.toString()}", status, ret);
+            Logger.Error($"cannot set property {name} to {rawValue.toString()}", status, ret);
         }
 
         internal static FREObject GetProperty(FREObject rawValue, string name) {
@@ -268,7 +263,7 @@ namespace TuaRua.FreSharp {
             var ret = Core.getProperty(rawValue, name, ref resultPtr);
             var status = (FreResultSharp) resultPtr;
             if (status == FreResultSharp.Ok) return ret;
-            Logger.Log($"cannot get property {name} of {rawValue.toString()}", status, ret);
+            Logger.Error($"cannot get property {name} of {rawValue.toString()}", status, ret);
             return FREObject.Zero;
         }
 
@@ -285,7 +280,7 @@ namespace TuaRua.FreSharp {
             var ret = Core.getActionScriptData(freContext, ref resultPtr);
             var status = (FreResultSharp) resultPtr;
             if (status == FreResultSharp.Ok) return ret;
-            Logger.Log("cannot get ActionScript data", status, ret);
+            Logger.Error("cannot get ActionScript data", status, ret);
             return FREObject.Zero;
         }
 
@@ -294,7 +289,7 @@ namespace TuaRua.FreSharp {
             Core.setActionScriptData(freContext, value, ref resultPtr);
             var status = (FreResultSharp) resultPtr;
             if (status == FreResultSharp.Ok) return;
-            Logger.Log("cannot set ActionScript data", status);
+            Logger.Error("cannot set ActionScript data", status);
         }
 
         internal static Dictionary<string, object> GetAsDictionary(FREObject rawValue) {