+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 2.3.0
+- Add default params to `FreArgException` 
+- Refactor `FREArray`
+- add optional `FREObject[] items` param to `FREArray`
+
 ### 2.2.0
 - Add `[DateTime]` support to `FREArray`
 - Change: `FREArray` now uses Vector.<Type> throughout